+++ r-package/tests/testthat/test-write_fare_structure.R
@@ -31,6 +31,4 @@ test_that("written structure is identical to original", {
 })
 
 # clean cache
-jar_dir <- tools::R_user_dir("r5r", which = "cache")
-f <- list.files(jar_dir, full.names = TRUE)
-unlink(f, recursive = TRUE)
+r5r::r5r_cache(delete_file = 'all')